saved changes

+++ mytaxi/UIImage+Tint.m
@@ -11,26 +11,29 @@
 @implementation UIImage (Tint)
 
 - (UIImage *)tintedImage:(UIColor *)color {
-    
     UIGraphicsBeginImageContextWithOptions(self.size, NO, [[UIScreen mainScreen] scale]);
     CGContextRef context = UIGraphicsGetCurrentContext();
     
     CGContextTranslateCTM(context, 0, self.size.height);
-    CGContextScaleCTM(context, 1.0, (CGFloat) -1.0);
+    CGContextScaleCTM(context, 1.0, -1.0);
     
     CGRect rect = CGRectMake(0, 0, self.size.width, self.size.height);
     
+  
     CGContextSetBlendMode(context, kCGBlendModeNormal);
     [[UIColor blackColor] setFill];
     CGContextFillRect(context, rect);
     
+
     CGContextSetBlendMode(context, kCGBlendModeNormal);
     CGContextDrawImage(context, rect, self.CGImage);
     
+  
     CGContextSetBlendMode(context, kCGBlendModeSourceAtop);
     [color setFill];
     CGContextFillRect(context, rect);
     
+
     CGContextSetBlendMode(context, kCGBlendModeDestinationIn);
     CGContextDrawImage(context, rect, self.CGImage);
     
@@ -38,7 +41,7 @@
     UIGraphicsEndImageContext();
     
     return tintedImage;
-    
 }
+
 
 @end
